Put Leveled into own module
