+++ lang/elaborator/src/normalizer/env.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 use derivative::Derivative;
 
 use pretty::DocAllocator;
-use syntax::ast::{Shift, ShiftRange};
+use syntax::ast::{Leveled, Shift, ShiftRange};
 
 use crate::normalizer::val::*;
 use printer::tokens::COMMA;
