+++ lang/syntax/src/ast/traits/mod.rs
@@ -1,11 +1,13 @@
 mod has_span;
 mod has_type;
+mod leveled;
 mod occurs;
 mod shift;
 pub mod subst;
 
 pub use has_span::*;
 pub use has_type::*;
+pub use leveled::*;
 pub use occurs::*;
 pub use shift::*;
 pub use subst::*;
