+++ lang/syntax/src/ctx/levels.rs
@@ -9,7 +9,7 @@ fn separated<I: IntoIterator<Item = String>>(s: &str, iter: I) -> String {
     vec.join(s)
 }
 
-use crate::common::*;
+use crate::{ast::Leveled, common::*};
 
 use super::def::*;
